requirements.txt
Flask==3.1.1
supabase
python-dotenv
